Initial setup
Adding Rubocop, setting Ruby version and setting dependency versions
