+++ .ruby-version
@@ -1,1 +1,1 @@
-3.1.4
+3.2.4
