+++ Gemfile
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 ruby File.read(".ruby-version").strip unless ENV["CI"]
 
 source "https://rubygems.org"
