--- content_block_presenter.gemspec
@@ -1,29 +1,0 @@
-
-lib = File.expand_path("../lib", __FILE__)
-$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "content_block_presenter/version"
-
-Gem::Specification.new do |spec|
-  spec.name          = "content_block_presenter"
-  spec.version       = ContentBlockPresenter::VERSION
-  spec.authors       = ["GOV.UK Dev"]
-  spec.email         = ["govuk-dev@digital.cabinet-office.gov.uk"]
-
-  spec.summary       = "A Ruby Gem that renders GOV.UK Publishing Content Blocks in a consistent way"
-  spec.homepage      = "https://github.com/alphagov/content_block_presenter"
-  spec.license       = "MIT"
-
-  spec.required_ruby_version = ">= 3.1"
-
-  # Specify which files should be added to the gem when it is released.
-  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
-  spec.files         = Dir.chdir(File.expand_path('..', __FILE__)) do
-    `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-  end
-  spec.bindir        = "exe"
-  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
-  spec.require_paths = ["lib"]
-
-  spec.add_development_dependency "rake", "~> 10.0"
-  spec.add_development_dependency "rspec", "~> 3.0"
-end
