--- lib/content_block_presenter.rb
@@ -1,6 +1,0 @@
-require "content_block_presenter/version"
-
-module ContentBlockPresenter
-  class Error < StandardError; end
-  # Your code goes here...
-end
