--- lib/content_block_presenter/version.rb
@@ -1,3 +1,0 @@
-module ContentBlockPresenter
-  VERSION = "0.1.0"
-end
