+++ lib/content_block_tools.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "content_block_tools/version"
+
+module ContentBlockTools
+  class Error < StandardError; end
+  # Your code goes here...
+end
