+++ lib/content_block_tools/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ContentBlockTools
+  VERSION = "0.1.0"
+end
