--- spec/content_block_presenter_spec.rb
@@ -1,9 +1,0 @@
-RSpec.describe ContentBlockPresenter do
-  it "has a version number" do
-    expect(ContentBlockPresenter::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
-  end
-end
