+++ spec/content_block_tools_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe ContentBlockTools do
+  it "has a version number" do
+    expect(ContentBlockTools::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
